# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HomeBaseCreate": {
        "additionalProperties": false,
        "properties": {
          "label": {
            "anyOf": [
              {
                "maxLength": 200,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "lat": {
            "anyOf": [
              {
                "maximum": 90.0,
                "minimum": -90.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Lat"
          },
          "lon": {
            "anyOf": [
              {
                "maximum": 180.0,
                "minimum": -180.0,
                "type": "number"
              },
              {
                "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
                "type": "string"
              }
            ],
            "title": "Lon"
          },
          "name": {
            "maxLength": 200,
            "minLength": 1,
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name",
          "lat",
          "lon"
        ],
        "title": "HomeBaseCreate",
        "type": "object"
      },
      "HomeBaseList": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/HomeBaseOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "HomeBaseList",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "Kitchen ERP",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/home-bases": {
      "get": {
        "operationId": "list_home_bases_api_v1_home_bases_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeBaseList"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "summary": "List Home Bases",
        "tags": [
          "geo",
          "home-bases"
        ]
      },
      "post": {
        "operationId": "create_home_base_api_v1_home_bases_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeBaseCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeBaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Home Base",
        "tags": [
          "geo",
          "home-bases"
        ]
      }
    },
    "/api/v1/home-bases/{home_base_id}": {
      "delete": {
        "operationId": "delete_home_base_api_v1_home_bases__home_base_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "home_base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Base Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Home Base",
        "tags": [
          "geo",
          "home-bases"
        ]
      },
      "get": {
        "operationId": "get_home_base_api_v1_home_bases__home_base_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "home_base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Base Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeBaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Home Base",
        "tags": [
          "geo",
          "home-bases"
        ]
      },
      "patch": {
        "operationId": "update_home_base_api_v1_home_bases__home_base_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "home_base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Base Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/HomeBaseUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HomeBaseOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Home Base",
        "tags": [
          "geo",
          "home-bases"
        ]
      }
    },
...
    "/api/v1/opening-hours/validate": {
      "post": {
        "operationId": "validate_opening_hours_api_v1_opening_hours_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OpeningHoursValidateIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OpeningHoursValidateOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Validate Opening Hours",
        "tags": [
          "geo",
          "opening-hours"
        ]
      }
    },
    "/api/v1/osm/adopt": {
      "post": {
        "operationId": "osm_adopt_api_v1_osm_adopt_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OsmAdoptIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Osm Adopt",
        "tags": [
          "geo",
          "osm"
        ]
      }
    },
    "/api/v1/osm/candidates": {
      "get": {
        "operationId": "osm_candidates_api_v1_osm_candidates_get",
        "parameters": [
          {
            "in": "query",
            "name": "home_base_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Home Base Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "radius_m",
            "required": true,
            "schema": {
              "maximum": 50000,
              "minimum": 100,
              "title": "Radius M",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OsmCandidateList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Osm Candidates",
        "tags": [
          "geo",
          "osm"
        ]
      }
    },
...
    "/api/v1/vendor-locations": {
      "get": {
        "operationId": "list_locations_api_v1_vendor_locations_get",
        "parameters": [
          {
            "in": "query",
            "name": "near",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 64,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Near"
            }
          },
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "chain",
                    "independent",
                    "market",
                    "stand"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "home_base_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Home Base Id"
            }
          },
          {
            "in": "query",
            "name": "vendor_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "uuid",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Vendor Id"
            }
          },
          {
            "in": "query",
            "name": "open_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Open At"
            }
          },
          {
            "in": "query",
            "name": "include_inactive",
            "required": false,
            "schema": {
              "default": false,
              "title": "Include Inactive",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationList"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Locations",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      },
      "post": {
        "operationId": "create_location_api_v1_vendor_locations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorLocationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Location",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      }
    },
    "/api/v1/vendor-locations/{location_id}": {
      "get": {
        "operationId": "get_location_api_v1_vendor_locations__location_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Location",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      },
      "patch": {
        "operationId": "update_location_api_v1_vendor_locations__location_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VendorLocationUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Location",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      }
    },
    "/api/v1/vendor-locations/{location_id}/activate": {
      "post": {
        "operationId": "activate_location_api_v1_vendor_locations__location_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Location",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      }
    },
    "/api/v1/vendor-locations/{location_id}/deactivate": {
      "post": {
        "operationId": "deactivate_location_api_v1_vendor_locations__location_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deactivate Location",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      }
    },
    "/api/v1/vendor-locations/{location_id}/is-open": {
      "get": {
        "operationId": "location_is_open_api_v1_vendor_locations__location_id__is_open_get",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "At"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IsOpenOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Location Is Open",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      }
    },
...
    "/api/v1/vendor-locations/{location_id}/refresh-osm": {
      "post": {
        "operationId": "refresh_location_osm_api_v1_vendor_locations__location_id__refresh_osm_post",
        "parameters": [
          {
            "in": "path",
            "name": "location_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Location Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorLocationDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Refresh Location Osm",
        "tags": [
          "geo",
          "vendor-locations"
        ]
      }
    },
...
    "/api/v1/vendors/{vendor_id}/activate": {
      "post": {
        "operationId": "activate_vendor_api_v1_vendors__vendor_id__activate_post",
        "parameters": [
          {
            "in": "path",
            "name": "vendor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vendor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Activate Vendor",
        "tags": [
          "geo",
          "vendors"
        ]
      }
    },
    "/api/v1/vendors/{vendor_id}/deactivate": {
      "post": {
        "operationId": "deactivate_vendor_api_v1_vendors__vendor_id__deactivate_post",
        "parameters": [
          {
            "in": "path",
            "name": "vendor_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Vendor Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VendorOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Deactivate Vendor",
        "tags": [
          "geo",
          "vendors"
        ]
      }
    }
  }
}